CloudSecurity/src/requirements.txt
functions-framework==3.*
requests
google-cloud-secret-manager==2.0.0